+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Debug logging in database submodule
+
+### Changed
+
+- Updated README.md
+
+### Fixed
+
+- Search query network list compilation
+
 ## [0.0.10] - 2022-03-17
 
 ### Changed